+++ cargo-workspaces/src/utils/changable.rs
@@ -7,10 +7,6 @@ use std::path::Path;
 #[derive(Debug, Parser)]
 pub struct ChangeOpt {
     // TODO: include_dirty
-    /// Include tags from merged branches
-    #[clap(long)]
-    pub include_merged_tags: bool,
-
     /// Always include targeted crates matched by glob even when there are no changes
     #[clap(long, value_name = "pattern")]
     pub force: Option<String>,