+++ cargo-workspaces/src/utils/changable.rs
@@ -7,17 +7,17 @@ use std::path::Path;
 #[derive(Debug, Parser)]
 pub struct ChangeOpt {
     // TODO: include_dirty
-    /// Include tags from merged branches
-    #[clap(long)]
-    pub include_merged_tags: bool,
-
     /// Always include targeted crates matched by glob even when there are no changes
-    #[clap(long, value_name = "pattern")]
+    #[clap(long, value_name = "PATTERN")]
     pub force: Option<String>,
 
     /// Ignore changes in files matched by glob
-    #[clap(long, value_name = "pattern")]
+    #[clap(long, value_name = "PATTERN")]
     pub ignore_changes: Option<String>,
+
+    /// Use this git reference instead of the last tag
+    #[clap(long, forbid_empty_values(true))]
+    pub since: Option<String>,
 }
 
 #[derive(Debug, Default)]
@@ -34,11 +34,22 @@ impl ChangeData {
             &["rev-list", "--tags", "--max-count=1"],
         )?;
 
-        let (_, count, _) = git(&metadata.workspace_root, &["rev-list", "--count", &sha])?;
+        if sha.is_empty() {
+            return Ok(Self {
+                count: "1".to_string(),
+                since: None,
+                ..Default::default()
+            });
+        }
+
+        let (_, count, _) = git(
+            &metadata.workspace_root,
+            &["rev-list", "--count", &format!("HEAD...{sha}")],
+        )?;
 
         let since = git(
             &metadata.workspace_root,
-            &["describe", "--exact-match", &sha],
+            &["describe", "--exact-match", "--tags", &sha],
         )
         .ok()
         .map(|x| x.1);
@@ -55,6 +66,7 @@ impl ChangeOpt {
     pub fn get_changed_pkgs(
         &self,
         metadata: &Metadata,
+        // Optional because there can be no tags
         since: &Option<String>,
         private: bool,
     ) -> Result<(Vec<Pkg>, Vec<Pkg>), Error> {