+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 * Allow `init` subcommand to use existing manifest
 * Improve error message when no public packages found with `list`
 * Added support for `2024` edition when creating package
+* Added `locked` flag to `publish` subcommand
+
+### Bug Fixes
+* Fixes issue with first publish to custom registries
+* Fixes issue with recognising the last tag with `changed`
 
 ## 0.3.6