+++ CHANGELOG.md
@@ -1,14 +1,21 @@
 # Changelog
 
-## 0.4.0
+## Unreleased
 
 ### BREAKING
 * Update MSRV to 1.78.0
+* Remove `include-merged-tags` option from `changed`, `version` & `publish` subcommands
 
 ### Enhancements
 * Allow `init` subcommand to use existing manifest
 * Improve error message when no public packages found with `list`
 * Added support for `2024` edition when creating package
+* Added `locked` flag to `publish` subcommand
+* Added `since` flag to `version` & `publish` subcommands
+
+### Bug Fixes
+* Fixes issue with first publish to custom registries
+* Fixes issue with recognising the last tag with `changed`
 
 ## 0.3.6